Added game periods and beginning of autonomous selection
Resolved #5, as the timer has been added and the period if the game (autonomous, teleop, climb) have been added as a subtitle.

Also added prioritized autonomous selection GUI elements. GUI still needs some graphical work, but the logic is there. Will also need a method to communicate the selection to the robot.